General fix-up for bugs and conversion to Maven.
