+++ README.md
@@ -1,4 +1,4 @@
 MineResetLite
 =============
 
-Lightweight implementation for automatic mine resets.
+Lightweight implementation for automatic mine resets, a continuation by VirtualByte.
