--- build.xml
@@ -1,22 +1,0 @@
-<project name="MineResetLite" default="build">
-    <property name="build.version" value="dev-unversioned"/>
-    <target name="build">
-        <mkdir dir="build"/>
-        <javac srcdir="src" destdir="build" debug="on">
-            <classpath>
-                <fileset dir="lib">
-                    <include name="*.jar"/>
-                </fileset>
-            </classpath>
-        </javac>
-        <copy todir="build">
-            <fileset dir="src">
-                <include name="*.yml"/>
-                <include name="**/*.properties"/>
-            </fileset>
-        </copy>
-        <replace file="build/plugin.yml" token="$version$" value="${build.version}"/>
-        <jar jarfile="MineResetLite.jar" basedir="build"/>
-        <delete dir="build"/>
-    </target>
-</project>
