--- src/config.yml
@@ -1,2 +1,0 @@
-#Set this value to false to disable the checking of updates
-check-for-updates: true
