+++ src/main/java/com/koletar/jj/mineresetlite/InvalidCommandArgumentsException.java
@@ -1,0 +1,9 @@
+package com.koletar.jj.mineresetlite;
+
+/**
+ * Thrown when a command's arguments were unsatisfactory.
+ *
+ * @author jjkoletar
+ */
+public class InvalidCommandArgumentsException extends Exception {
+}
