+++ src/main/resources/phrases.properties
@@ -1,0 +1,70 @@
+mineList=&dMines: %0%
+firstPointSet=&dYour first point has been set.
+secondPointSet=&dYour second point has been set.
+crossWorldSelection=&4Your selection points must be in the same world.
+emptySelection=&4Make a selection first.
+nameInUse=&4The name &c%0% &4isn't original enough.
+mineCreated=&dMine &c%0% &dsaved.
+tooManyMines=&4Too many mines matched. Try being more specific.
+noMinesMatched=&4No mine was matched.
+mineInfoName=&dMine: &c%0%
+mineInfoWorld=&dWorld: &c%0%
+mineInfoComposition=&dComposition: &c%0%
+mineInfoResetDelay=&dResets every: &c%0% &dminute(s).
+mineInfoWarningTimes=&dReset warnings: &c%0% &dminute(s).
+mineInfoSurface=&dSurface Block: &c%0%
+mineInfoFillMode=&dFill mode &cis &denabled.
+mineInfoTimeUntilReset=&dResets in: &cless than %0% &dminute(s).
+mineInfoSilence=&dResets silently: &c%0%
+unknownBlock=&4Unknown block.
+mineCompositionSet=&c%0% &dis now %1%% %2%. Mine is now &c%3%% &dempty.
+badPercentage=&4Invalid percentage.
+insaneCompositionChange=&4A mine's composition cannot add up to be greater than 100%.
+mineResetBroadcast=&f[&4MRL&f] &c%0% &dhas been reset by &c%1%&d!
+mineAutoResetBroadcast=&f[&4MRL&f] &c%0% &dhas been reset automatically!
+mineWarningBroadcast=&f[&4MRL&f] &c%0% &dwill reset in &5%1% &dminute(s)!
+notABlock=&4Please select a block.
+blockNotInMine=&c%0% &4doesn't contain any &c%1%&4.
+blockRemovedFromMine=&c%0% &dno longer contains &c%1%&d. Mine is now &c%2%% &dempty.
+badResetDelay=&4Invalid delay.
+badWarningList=&4Invalid list of warning times.
+warningListCleared=&c%0%&d's warning list has been cleared.
+warningListSet=&c%0%&d's warning list has been set.
+resetDelaySet=&c%0% &dwill reset every &c%1% &dminute(s).
+resetDelayCleared=&c%0% &dwill no longer automatically reset.
+unknownFlag=&4Unknown flag. Use /mrl ? set for a list.
+mineErased=&c%0% &dhas been erased.
+unknownCommand=&4Unknown command.
+invalidArguments=&4Invalid arguments.
+invalidArgsUsage=&4/mrl %0% %1%
+notAPlayer=&4You must be a player to use this command.
+noPermission=&4You don't have permission to do that.
+helpUsage=&e/mrl %0% %1%
+helpDesc=&7 - %0%
+aboutTitle=&f[&4MineResetLite&f]
+aboutAuthors=&bby jjkoletar
+aboutVersion=&dVersion: &c%0%
+surfaceBlockSet=&c%0%&d's surface block has been set.
+surfaceBlockCleared=&c%0%&d's surface block has been cleared.
+updateWarning1=&f[&4MineResetLite&f] &dAn update has been released. Please update soon at:
+updateWarning2=&dhttp://dev.bukkit.org/server-mods/mineresetlite/
+criticalUpdateWarningDecoration=&4===========
+criticalUpdateWarning=&dThis update is CRITICAL. It may address urgent security flaws or exploits.
+fillModeEnabled=&dFill mode enabled.
+fillModeDisabled=&dFill mode disabled.
+invalidFillMode=&4Invalid fill mode value.
+mineIsNowSilent=&c%0% &dnow resets silently.
+mineIsNoLongerSilent=&c%0% &dno longer resets silently.
+badBoolean=&4Unknown value. Valid values: true, false
+rescheduled=&dAll mines rescheduled.
+signReaderKey=[MRL]
+signReaderNoPerm=&4[No Permission]
+signReaderResetTime=!time
+signReaderReset=!reset
+signUnknownMine=&4[Unknown Mine]
+signReset=&f[&4Reset&f]
+signMineName=&1%0%
+resetSign1=&c%0%
+resetSign2=&dwill reset in
+resetSign3=&d%0%
+console=Console
