+++ src/main/resources/phrases_es.properties
@@ -1,0 +1,65 @@
+mineList=&dMinas: %0%
+firstPointSet=&dTu primer punto ha sido seleccionado.
+secondPointSet=&dTu segundo punto ha sido seleccionado.
+crossWorldSelection=&4Tus puntos de seleccion deben estar en el mismo mundo.
+emptySelection=&4Has primero una seleccion.
+nameInUse=&4El nombre &c%0% &4esta en uso.
+mineCreated=&dMina &c%0% &dguardada.
+tooManyMines=&4Muchas minas encontradas. Intentalo siendo mas especifico.
+noMinesMatched=&4No se encuentro la mina.
+mineInfoName=&dMina: &c%0%
+mineInfoWorld=&dMundo: &c%0%
+mineInfoComposition=&dComposicion: &c%0%
+mineInfoResetDelay=&dReinicios cada: &c%0% &dminuto(s).
+mineInfoWarningTimes=&dAvisos de reinicio: &c%0% &dminuto(s).
+mineInfoSurface=&dSuperficie del bloque: &c%0%
+mineInfoFillMode=&dModo de relleno &cesta &dactivado.
+unknownBlock=&4Bloque desconocido.
+mineCompositionSet=&c%0% &des ahora %1%% %2%.
+badPercentage=&4Porcentage invalido.
+insaneCompositionChange=&4La composicion de una mina no puede superar el 100%.
+mineResetBroadcast=&f[&4MRL&f] &c%0% &dha sido reiniciada por &c%1%&d!
+mineAutoResetBroadcast=&f[&4MRL&f] &c%0% &dha sido reiniciada automaticamente!
+mineWarningBroadcast=&f[&4MRL&f] &c%0% &dse reiniciara en &5%1% &dminuto(s)!
+notABlock=&4Por favor selecciona un bloque.
+blockNotInMine=&c%0% &4no contiene ningun &c%1%&4.
+blockRemovedFromMine=&c%0% &dya no contiene &c%1%&d.
+badResetDelay=&4retardo no valido.
+badWarningList=&4Lista de tiempos de alerta invalida.
+warningListCleared=&c%0%&d su lista de alertas ha sido borrada.
+warningListSet=&c%0%&d su lista de alerta se ha establecido.
+resetDelaySet=&c%0% &dse reiniciara cada &c%1% &dminuto(s).
+resetDelayCleared=&c%0% &dya no se reinicia automaticamente.
+unknownFlag=&4Flag desconocida. Utilice /mrl ? para conocer la lista de comandos.
+mineErased=&c%0% &dha sido borrada.
+unknownCommand=&4Comando desconocido.
+invalidArguments=&4Argumentos invalidos.
+invalidArgsUsage=&4/mrl %0% %1%
+notAPlayer=&4Necesitas ser un jugador para usar este comando.
+noPermission=&4No tienes permisos para hacer esto.
+helpUsage=&e/mrl %0% %1%
+helpDesc=&7 - %0%
+aboutTitle=&f[&4MineResetLite&f]
+aboutAuthors=&bby jjkoletar
+aboutVersion=&dVersion: &c%0%
+surfaceBlockSet=&c%0%&d su bloque de superficie se ha establecido.
+surfaceBlockCleared=&c%0%&d su bloque de superficie ha sido limpiada.
+updateWarning1=&f[&4MineResetLite&f] &dHay una actualizacion. Por favor actualice pronto en:
+updateWarning2=&dhttp://dev.bukkit.org/server-mods/mineresetlite/
+criticalUpdateWarningDecoration=&4===========
+criticalUpdateWarning=&dEsta actualizacion es CRITICA. Puede tratar defectos de seguridad urgentes o exploits.
+fillModeEnabled=&dModo de rellenado habilitado.
+fillModeDisabled=&dModo de rellenado deshabilitado.
+invalidFillMode=&4Valor de rellenado invalido.
+rescheduled=&dTodas las minas reprogramadas.
+signReaderKey=[MRL]
+signReaderNoPerm=&4[Sin permisos]
+signReaderResetTime=!tiempo
+signReaderReset=!reinicio
+signUnknownMine=&4[Mina desconocida]
+signReset=&f[&4Reinicio&f]
+signMineName=&1%0%
+resetSign1=&c%0%
+resetSign2=&dse reiniciara en
+resetSign3=&d%0%
+console=Consola
