+++ src/main/resources/phrases_fr.properties
@@ -1,0 +1,52 @@
+mineList=&dMines: %0%
+firstPointSet=&dVotre premier point a été réglé.
+secondPointSet=&dVotre deuxième point a été réglé.
+crossWorldSelection=&4Vos points de sélection doivent être dans le même monde.
+emptySelection=&4Faites vos selection en premier.
+nameInUse=&4The name &c%0% &4est déjà utilisé
+mineCreated=&dMine &c%0% &dsauvegardé.
+tooManyMines=&4Trop de mines assortis. Essayez d'être plus précis.
+noMinesMatched=&4Aucune mine a été apparié.
+mineInfoName=&dMine: &c%0%
+mineInfoWorld=&dWorld: &c%0%
+mineInfoComposition=&dComposition: &c%0%
+mineInfoResetDelay=&dRéinitialise tous: &c%0% &dminute(s).
+mineInfoWarningTimes=&dréinitialiser avertissements: &c%0% &dminute(s).
+mineInfoSurface=&dBloc de surface: &c%0%
+mineInfoFillMode=&dmode de remplissage &cest &dactivé.
+mineInfoTimeUntilReset=&dRéinitialise dans: &cmoins de %0% &dminute(s).
+mineInfoSilence=&dRéinitialise en silence: &c%0%
+unknownBlock=&4bloc inconnu.
+mineCompositionSet=&c%0% &dest maintenant %1%% %2%.
+badPercentage=&4Pourcentage invalide.
+insaneCompositionChange=&4A La composition de la mine ne peut pas être supérieure à 100%.
+mineResetBroadcast=&f[&4MRL&f] &c%0% &da été réinitialisée par &c%1%&d!
+mineAutoResetBroadcast=&f[&4MRL&f] &c%0% &da été réinitialisée automatiquement!
+mineWarningBroadcast=&f[&4MRL&f] &c%0% &dva se réinitialiser dans &5%1% &dminute(s)!
+notABlock=&4S'il vous plaît, séléctionner un bloc.
+blockNotInMine=&c%0% &4ne contient aucun &c%1%&4.
+blockRemovedFromMine=&c%0% &dne contient plus &c%1%&d.
+badResetDelay=&4Délai invalide.
+badWarningList=&4Liste non valide du temps d'avertissement.
+warningListCleared=&c%0%&d's liste d'alerte a été effacée.
+warningListSet=&c%0%&d's liste d'alerte a été défini.
+resetDelaySet=&c%0% &dsera réinitialiser tous les &c%1% &dminute(s).
+resetDelayCleared=&c%0% &dne sera plus automatiquement réinitialisé.
+unknownFlag=&4drapeau Inconnu. Utilisez /mrl ? définir une liste.
+mineErased=&c%0% &da été effacée
+unknownCommand=&4Commande inconnue.
+invalidArguments=&4Arguments invalides.
+invalidArgsUsage=&4/mrl %0% %1%
+notAPlayer=&4Vous devez être un joueur pour utiliser cette commande!
+noPermission=&4Vous n'avez pas la permission de faire cela!
+helpUsage=&e/mrl %0% %1%
+helpDesc=&7 - %0%
+aboutTitle=&f[&4MineResetLite&f]
+aboutAuthors=&bby jjkoletar
+aboutVersion=&dVersion: &c%0%
+surfaceBlockSet=&c%0%&d's Bloc de surface a été défini.
+surfaceBlockCleared=&c%0%&d's bloc de surface a été effacée.
+updateWarning1=&f[&4MineResetLite&f] &dUne mise à jour a été publiée. S'il vous plaît mettre à jour rapidement à:
+updateWarning2=&dhttp://dev.bukkit.org/server-mods/mineresetlite/
+criticalUpdateWarningDecoration=&4===========
+criticalUpdateWarning=&dCette mise à jour est critique. Il peut répondre à des failles de sécurité urgentes ou des exploits.
