+++ src/main/resources/phrases_pl.properties
@@ -1,0 +1,71 @@
+## tlumaczenie by areqpl/gg:1930279
+mineList=&dObszary: %0%
+firstPointSet=&dTwoj pierwszy punkt zostal ustawiony.
+secondPointSet=&dTwoj drugi punkt zostal ustawiony.
+crossWorldSelection=&4Twoje punkty zaznaczenia musza byc w tym samym swiecie.
+emptySelection=&4Najpierw zaznacz region.
+nameInUse=&4Nazwa &c%0% &4nie jest unikalna.
+mineCreated=&dObszar &c%0% &dzapisany.
+tooManyMines=&4Zbyt wiele podobnych wynikow. Postaraj sie byc bardziej konretnym.
+noMinesMatched=&4Nie znaleziono takiego obszaru.
+mineInfoName=&dObszar: &c%0%
+mineInfoWorld=&dSwiat: &c%0%
+mineInfoComposition=&dKompozycja: &c%0%
+mineInfoResetDelay=&dReset co: &c%0% &dminut.
+mineInfoWarningTimes=&dOstrzezenia co: &c%0% &dminut.
+mineInfoSurface=&dWarstwa powierzchniowa: &c%0%
+mineInfoFillMode=&dWypelnianie &cjest &dwlaczone.
+mineInfoTimeUntilReset=&dReset w czasie: &cmniejszym niz %0% &dminut.
+mineInfoSilence=&dCichy reset: &c%0%
+unknownBlock=&4Nieznany blok.
+mineCompositionSet=&c%0% &djest teraz %1%% %2%.
+badPercentage=&4Podales niepoprawny procent.
+insaneCompositionChange=&4Nie mozesz przekroczyc wartosci 100% podczas tworzenia kompozycji obszaru.
+mineResetBroadcast=&f[&4MRL&f] &dObszar &c%0% &dzostal zresetowany przez &c%1%&d!
+mineAutoResetBroadcast=&f[&4MRL&f] &dObszar &c%0% &dzostal automatycznie zresetowany!
+mineWarningBroadcast=&f[&4MRL&f] &dObszar &c%0% &dzostanie zresetowany za &5%1% &dminut!
+notABlock=&4Prosze wybrac blok.
+blockNotInMine=&c%0% &4nie zawieral blokow &c%1%&4.
+blockRemovedFromMine=&c%0% &dnie zawiera juz blokow &c%1%&d.
+badResetDelay=&4Podales nieprawidlowy czas.
+badWarningList=&4Podales nieprawidlowa liste czasow ostrzezen.
+warningListCleared=&dLista ostrzezen obszaru &c%0%&d zostala wyzerowana.
+warningListSet=&dLista ostrzezen obszaru &c%0%&d zostala ustawiona.
+resetDelaySet=&c%0% &dbedzie resetowac sie co &c%1% &dminut.
+resetDelayCleared=&c%0% &dnie bedzie juz sie resetowac.
+unknownFlag=&4Nieznana flaga. Wpisz /mrl ? set aby otrzymac ich liste.
+mineErased=&c%0% &dzostal usuniety.
+unknownCommand=&4Nieznana komenda.
+invalidArguments=&4Nieprawidlowy argument.
+invalidArgsUsage=&4/mrl %0% %1%
+notAPlayer=&4Musisz byc graczem zeby uzyc tej komendy.
+noPermission=&4Nie masz uprawnien do tej czynnosci.
+helpUsage=&e/mrl %0% %1%
+helpDesc=&7 - %0%
+aboutTitle=&f[&4MineResetLite&f]
+aboutAuthors=&bplugin stworzony przez jjkoletar
+aboutVersion=&dWersja: &c%0%
+surfaceBlockSet=&dBlok powierzchniowy obszaru &c%0%&d zostal ustawiony.
+surfaceBlockCleared=&dBlok powierzchniowy obszaru &c%0%&d zostal wyzerowany.
+updateWarning1=&f[&4MineResetLite&f] &dDostepna jest nowa aktualizacja. Zaktualizuj pobierajac plugin:
+updateWarning2=&dhttp://dev.bukkit.org/server-mods/mineresetlite/
+criticalUpdateWarningDecoration=&4===========
+criticalUpdateWarning=&dTa aktualizacja jest KRYTYCZNA. Moze zawierac bardzo wazne latki bezpieczenstwa.
+fillModeEnabled=&dWypelnianie wlaczone.
+fillModeDisabled=&dWypelnianie wylaczone.
+invalidFillMode=&4Nieprawidlowa wartosc wypelniania.
+mineIsNowSilent=&c%0% &dod teraz bedzie resetowac sie po cichu.
+mineIsNoLongerSilent=&c%0% &dod teraz bedzie oglaszac reset.
+badBoolean=&4Nieznana wartosc. Prawidlowe wartosci: true, false
+rescheduled=&dWszystkie obszary zostaly zsynchronizowane.
+signReaderKey=[MRL]
+signReaderNoPerm=&4[Brak uprawnien]
+signReaderResetTime=!czas
+signReaderReset=!reset
+signUnknownMine=&4[Nieznany Obszar]
+signReset=&f[&4Reset&f]
+signMineName=&1%0%
+resetSign1=&c%0%
+resetSign2=&dreset za
+resetSign3=&d%0%
+console=Konsola
