--- src/phrases_da.properties
@@ -1,65 +1,0 @@
-mineList=&dMiner: %0%
-firstPointSet=&dDit første punkt er markeret.
-secondPointSet=&dDit andet punkt er markeret.
-crossWorldSelection=&4Dine punkter skal være i samme verden.
-emptySelection=&4Du skal først markere et område.
-nameInUse=&4Navnet &c%0% &4er ikke originalt nok.
-mineCreated=&dMine &c%0% &dgemt.
-tooManyMines=&4For mange miner fundet. Prøv noget mere specifikt.
-noMinesMatched=&4Ingen miner fundet.
-mineInfoName=&dMine: &c%0%
-mineInfoWorld=&dVerden: &c%0%
-mineInfoComposition=&dSammensætning: &c%0%
-mineInfoResetDelay=&dGenstarter hver: &c%0% &dminut(er).
-mineInfoWarningTimes=&dGenstarts advarsler: &c%0% &dminut(er).
-mineInfoSurface=&dOverflade Blok: &c%0%
-mineInfoFillMode=&dFyldnings tilstand &cis &dAktiveret.
-unknownBlock=&4Ukendt Blok.
-mineCompositionSet=&c%0% &der nu %1%% %2%.
-badPercentage=&4Forkert Procent.
-insaneCompositionChange=&4En mines sammensætning kan ikke være mere end 100%.
-mineResetBroadcast=&f[&4MRL&f] &c%0% &der blevet genstartet af &c%1%&d!
-mineAutoResetBroadcast=&f[&4MRL&f] &c%0% &der blevet genstartet automatisk!
-mineWarningBroadcast=&f[&4MRL&f] &c%0% &dvil genstarte om &5%1% &dminut(er)!
-notABlock=&4Vælg venligst en blok.
-blockNotInMine=&c%0% &4Indeholder ingen &c%1%&4.
-blockRemovedFromMine=&c%0% &dIndeholder ikke længere &c%1%&d.
-badResetDelay=&4Forkert forsinkelse.
-badWarningList=&4Forkert liste af advarsels gange.
-warningListCleared=&c%0%&d's advarsels liste er blevet ryddet.
-warningListSet=&c%0%&d's advarsels liste er blevet sat.
-resetDelaySet=&c%0% &dVil genstarte hver &c%1% &dminut(er).
-resetDelayCleared=&c%0% &dVil ikke længere automatisk genstarte.
-unknownFlag=&4Ukendt. brug /mrl ? set for en liste.
-mineErased=&c%0% &der blevet slettet.
-unknownCommand=&4Ukendt kommando.
-invalidArguments=&4Forkerte argumenter.
-invalidArgsUsage=&4/mrl %0% %1%
-notAPlayer=&Du skal være en spiller for at bruge denne kommando.
-noPermission=&4Du har ikke tilladelse til at gøre det.
-helpUsage=&e/mrl %0% %1%
-helpDesc=&7 - %0%
-aboutTitle=&f[&4MineResetLite&f]
-aboutAuthors=&baf jjkoletar
-aboutVersion=&dVersion: &c%0%
-surfaceBlockSet=&c%0%&d's Overflade blok er blevet valgt.
-surfaceBlockCleared=&c%0%&d's overflade blok er blevet ryddet.
-updateWarning1=&f[&4MineResetLite&f] &dEn opdatering er klar, opdater venligst snart her:
-updateWarning2=&dhttp://dev.bukkit.org/server-mods/mineresetlite/
-criticalUpdateWarningDecoration=&4===========
-criticalUpdateWarning=&dDenne obdatering er VIGTIG. Den kan stoppe sikkerheds brud og udnyttelser.
-fillModeEnabled=&dFyldnings tilstand aktiveret.
-fillModeDisabled=&dFyldnings tilstand deaktiveret.
-invalidFillMode=&4Forkert fyldnings tilstands værdi.
-rescheduled=&dAlle miner omlagt.
-signReaderKey=[MRL]
-signReaderNoPerm=&4[Ingen Tilladelse]
-signReaderResetTime=!tid
-signReaderReset=!genstart
-signUnknownMine=&4[Ukendt Mine]
-signReset=&f[&4Genstart&f]
-signMineName=&1%0%
-resetSign1=&c%0%
-resetSign2=&dvil genstarte om
-resetSign3=&d%0%
-console=Konsol
