--- src/phrases_it.properties
@@ -1,70 +1,0 @@
-mineList=&dMine: %0%
-firstPointSet=&dPrimo punto settato.
-secondPointSet=&dSecondo punto settato.
-crossWorldSelection=&4I tuoi punti devono essere nello stesso mondo.
-emptySelection=&4Seleziona un punto prima.
-nameInUse=&4Il nome &c%0% &4Non Ë originale.
-mineCreated=&dMina &c%0% &dSalvata.
-tooManyMines=&4Troppe Mine Uguali. Sii piu' specifico.
-noMinesMatched=&4Nessuna mina corrisponde.
-mineInfoName=&dMina: &c%0%
-mineInfoWorld=&dWorld: &c%0%
-mineInfoComposition=&dComposizione: &c%0%
-mineInfoResetDelay=&dResets ogni: &c%0% &dMinuti(s).
-mineInfoWarningTimes=&dAttenzione Reset: &c%0% &dMinuti(s).
-mineInfoSurface=&dBlocco di superficie: &c%0%
-mineInfoFillMode=&dModalita' riempimento &ce' &dAbilitata.
-mineInfoTimeUntilReset=&dResets in: &Meno di %0% &dminuti(s).
-mineInfoSilence=&dResets silenzioso: &c%0%
-unknownBlock=&4Blocco sconosciuto.
-mineCompositionSet=&c%0% &dOra e' %1%% %2%.
-badPercentage=&4Percentuale invalida.
-insaneCompositionChange=&4La composizione di una mina non puo' essere piu' vantaggiosa del 100%.
-mineResetBroadcast=&f[&4MRL&f] &c%0% &dSono stati resettati da &c%1%&d!
-mineAutoResetBroadcast=&f[&4MRL&f] &c%0% &dSono stati resettati automaticamente!
-mineWarningBroadcast=&f[&4MRL&f] &c%0% &dSaranno resettati in &5%1% &dminuti(s)!
-notABlock=&4Perfevore seleziona un blocco.
-blockNotInMine=&c%0% &4Non contiene nessun &c%1%&4.
-blockRemovedFromMine=&c%0% &dNon lo contiene piu' &c%1%&d.
-badResetDelay=&4Ritardo invalido.
-badWarningList=&4Lista warnings invalida.
-warningListCleared=&c%0%&d Lista warning e' stata resettata.
-warningListSet=&c%0%&d Lista warning settata.
-resetDelaySet=&c%0% &dSara' resettata ogni &c%1% &dminuti(s).
-resetDelayCleared=&c%0% &dNon sara' piu' resettata automaticamente.
-unknownFlag=&4Flag sconosciuta. Usa /mrl ? Per la lista.
-mineErased=&c%0% &dSono stati cancellati.
-unknownCommand=&4Comando sconosciuto.
-invalidArguments=&4Argomento invalido.
-invalidArgsUsage=&4/mrl %0% %1%
-notAPlayer=&4Devi essere un player per usare il comando.
-noPermission=&4Non hai i permessi per usare il comando.
-helpUsage=&e/mrl %0% %1%
-helpDesc=&7 - %0%
-aboutTitle=&f[&4MineResetLite&f]
-aboutAuthors=&bby jjkoletar
-aboutVersion=&dVersione: &c%0%
-surfaceBlockSet=&c%0%&d Blocchi di superficie sono stati settati.
-surfaceBlockCleared=&c%0%&d Blocchi di superficie sono stati cancellati.
-updateWarning1=&f[&4MineResetLite&f] &dUn aggiornamento e' disponibile. Perfevore aggiorna presto a:
-updateWarning2=&dhttp://dev.bukkit.org/server-mods/mineresetlite/
-criticalUpdateWarningDecoration=&4===========
-criticalUpdateWarning=&dQuesto aggiornamento e' CRITICO. Puo' contenere aggiormaneti importanti.
-fillModeEnabled=&dFill mode Abilitata.
-fillModeDisabled=&dFill mode disabilitata.
-invalidFillMode=&4Valore dato alla Fill Mode Invalido.
-mineIsNowSilent=&c%0% &dOra si resetta silenziosamente.
-mineIsNoLongerSilent=&c%0% &dNon si resetta piu' silenziosamente.
-badBoolean=&4Valore sconosciuto. Argomenti validi: true, false
-rescheduled=&dTutte le mine riprogrammate.
-signReaderKey=[MRL]
-signReaderNoPerm=&4[No Permission]
-signReaderResetTime=!time
-signReaderReset=!reset
-signUnknownMine=&4[Unknown Mine]
-signReset=&f[&4Reset&f]
-signMineName=&1%0%
-resetSign1=&c%0%
-resetSign2=&dSara' resettato in
-resetSign3=&d%0%
-console=Console
