--- src/phrases_nl.properties
@@ -1,70 +1,0 @@
-mineList=&dMijnen: %0%
-firstPointSet=&dJe eerste punt is gezet.
-secondPointSet=&dJe tweede punt is gezet.
-crossWorldSelection=&4Je selectie punten moeten zich in de zelfde wereld bevinden.
-emptySelection=&4Maak eerst een selectie.
-nameInUse=&4De naam &c%0% &4is niet origineel genoeg.
-mineCreated=&dMijn &c%0% &dsaved.
-tooManyMines=&4Te veel mijnen gevonden. Probeer meer specifiek te zijn.
-noMinesMatched=&4Geen enkele mijn gevonden.
-mineInfoName=&dMijn: &c%0%
-mineInfoWorld=&dWereld: &c%0%
-mineInfoComposition=&dCompositie: &c%0%
-mineInfoResetDelay=&dreset elke: &c%0% &dminuut.
-mineInfoWarningTimes=&dReset Waarschuwingen: &c%0% &dminuut.
-mineInfoSurface=&dOppervlak blok: &c%0%
-mineInfoFillMode=&dVolle modus &cis &daangezet.
-mineInfoTimeUntilReset=&dReset in: &cminder dan %0% &dminuut.
-mineInfoSilence=&dReset zachtjes: &c%0%
-unknownBlock=&4Onbekend blok.
-mineCompositionSet=&c%0% &dis nu %1%% %2%.
-badPercentage=&4Ongeldig percentage.
-insaneCompositionChange=&4De compositie van een mijn mag niet groter dan 100% zijn.
-mineResetBroadcast=&f[&4MRL&f] &c%0% &dIs gereset door&c%1%&d!
-mineAutoResetBroadcast=&f[&4MRL&f] &c%0% &dIs automatisch gereset!
-mineWarningBroadcast=&f[&4MRL&f] &c%0% &zal resetten in&5%1% &dminuut!
-notABlock=&4Selecteer een blok.
-blockNotInMine=&c%0% &4Bevat geen&c%1%&4.
-blockRemovedFromMine=&c%0% &dBevat niet langer&c%1%&d.
-badResetDelay=&4Ongeldige vertraging.
-badWarningList=&4Ongeldige lijst van waarschuwings tijden.
-warningListCleared=&c%0%&d's waarschuwings lijst is gewist.
-warningListSet=&c%0%&d's Waarschuwings lijst is ingesteld.
-resetDelaySet=&c%0% &dzal elke &c%1% &dminuut reseten.
-resetDelayCleared=&c%0% &dzal niet langer automatisch reseten.
-unknownFlag=&4Onbekende vlag. Gebruik /mrl ? set voor een lijst.
-mineErased=&c%0% &dIs gewist.
-unknownCommand=&4Onbekend commando.
-invalidArguments=&4Ongeldige woorden.
-invalidArgsUsage=&4/mrl %0% %1%
-notAPlayer=&4Je moet een speler zijn om dit commando te gebruiken.
-noPermission=&4Je hebt geen toestemming om dat te doen.
-helpUsage=&e/mrl %0% %1%
-helpDesc=&7 - %0%
-aboutTitle=&f[&4MineResetLite&f]
-aboutAuthors=&bgemaakt door jjkoletar, vertaald door banaantjuh
-aboutVersion=&dVersion: &c%0%
-surfaceBlockSet=&c%0%&d's oppervlakte blokken zijn ingesteld.
-surfaceBlockCleared=&c%0%&d's oppervlakte blokken zijn gewist.
-updateWarning1=&f[&4MineResetLite&f] &dEen update is vrij gegeven. Update zo snel mogelijk op:
-updateWarning2=&dhttp://dev.bukkit.org/server-mods/mineresetlite/
-criticalUpdateWarningDecoration=&4===========
-criticalUpdateWarning=&dDeze update is niet helemaal veilig. Het kan beveiligingsfouten bevatten.
-fillModeEnabled=&dVul modus aangezet.
-fillModeDisabled=&dVul modus uitgezet.
-invalidFillMode=&4Ongeldige vul modus waarde.
-mineIsNowSilent=&c%0% &dreset nu zachtjes.
-mineIsNoLongerSilent=&c%0% &dreset niet langer zachtjes.
-badBoolean=&4Onbekende waarde. Geldige waardes: True, False
-rescheduled=&dAlles mijnen zijn hergeschikt.
-signReaderKey=[MRL]
-signReaderNoPerm=&4[Geen toestemming]
-signReaderResetTime=!time
-signReaderReset=!reset
-signUnknownMine=&4[Onbekende mijn]
-signReset=&f[&4Reset&f]
-signMineName=&1%0%
-resetSign1=&c%0%
-resetSign2=&dZal resetten in
-resetSign3=&d%0%
-console=Console
