--- src/phrases_pt.properties
@@ -1,65 +1,0 @@
-mineList=&dMinas: %0%
-firstPointSet=&dSeu primeiro ponto foi selecionado.
-secondPointSet=&dSeu segundo ponto foi selecionado.
-crossWorldSelection=&4Seus pontos de seleção devem estar no mesmo mundo.
-emptySelection=&4Faça uma seleção primeiro.
-nameInUse=&4O nome &c%0% &4já está sendo usado.
-mineCreated=&dMina &c%0% &dsalva.
-tooManyMines=&4Muitas minas encontradas. Tente ser mais específico.
-noMinesMatched=&4Nenhuma mina foi encontrada.
-mineInfoName=&dMina: &c%0%
-mineInfoWorld=&dMundo: &c%0%
-mineInfoComposition=&dComposição: &c%0%
-mineInfoResetDelay=&dResetado em: &c%0% &dminuto(s).
-mineInfoWarningTimes=&dAvisos resetados em: &c%0% &dminuto(s).
-mineInfoSurface=&dBloco da superfície: &c%0%
-mineInfoFillMode=&dModo de preenchimento &cestá &dativado.
-unknownBlock=&4Bloco desconhecido.
-mineCompositionSet=&c%0% &dé agora %1%% %2%.
-badPercentage=&4Porcetagem inválida.
-insaneCompositionChange=&4A composição da mina não pode ser maior do que 100%.
-mineResetBroadcast=&f[&4MRL&f] &c%0% &dfoi resetada por &c%1%&d!
-mineAutoResetBroadcast=&f[&4MRL&f] &c%0% &dfoi resetada automaticamente!
-mineWarningBroadcast=&f[&4MRL&f] &c%0% &dserá resetada em &5%1% &dminuto(s)!
-notABlock=&4Por favor, selecione um bloco.
-blockNotInMine=&c%0% &4não contém nenhum &c%1%&4.
-blockRemovedFromMine=&c%0% &djá não contém &c%1%&d.
-badResetDelay=&4Delay inválido.
-badWarningList=&4Lista de horários de alertas inválido.
-warningListCleared=&c%0%&d lista de alerta foi deletada.
-warningListSet=&c%0%&d lista de alerta foi criada.
-resetDelaySet=&c%0% &dserá sempre resetada em &c%1% &dminuto(s).
-resetDelayCleared=&c%0% &dnão será mais resetada automaticamente.
-unknownFlag=&4Flag desconhecida. Use /mrl ? set para ver a lista.
-mineErased=&c%0% &dfoi deletada.
-unknownCommand=&4Comando desconhecido.
-invalidArguments=&4Argumentos inválidos.
-invalidArgsUsage=&4/mrl %0% %1%
-notAPlayer=&4Você deve ser um jogador para usar este comando.
-noPermission=&4Você não tem permissão para fazer isso.
-helpUsage=&e/mrl %0% %1%
-helpDesc=&7 - %0%
-aboutTitle=&f[&4MineResetLite&f]&f
-aboutAuthors=&bpor jjkoletar
-aboutVersion=&dVersão: &c%0%
-surfaceBlockSet=&c%0%&dbloco da superfície foi definido.
-surfaceBlockCleared=&c%0%&dbloco da superfície foi deletado.
-updateWarning1=&f[&4MRL&f] &dUma atualização foi liberada. Por favor atualize no:
-updateWarning2=&dhttp://dev.bukkit.org/server-mods/mineresetlite/
-criticalUpdateWarningDecoration=&4===========
-criticalUpdateWarning=&dEsta atualização é crítica. Pode resolver bugs urgentes.
-fillModeEnabled=&dModo preencher ativado.
-fillModeDisabled=&dModo preencher desativado.
-invalidFillMode=&4Valor do modo preencher inválido.
-rescheduled=&dTodas as minas foram remarcadas.
-signReaderKey=[MRL]
-signReaderNoPerm=&4[Sem Permissão]
-signReaderResetTime=!tempo
-signReaderReset=!resetar
-signUnknownMine=&4[Mina desconhecida]
-signReset=&f[&4Resetar&f]
-signMineName=&1%0%
-resetSign1=&c%0%
-resetSign2=&dserá resetada em
-resetSign3=&d%0%
-console=Console
