--- src/plugin.yml
@@ -1,10 +1,0 @@
-name: MineResetLite
-version: $version$
-main: com.koletar.jj.mineresetlite.MineResetLite
-author: jjkoletar
-description: A lightweight solution for Mine management.
-softdepend: [WorldEdit, Multiverse-Core]
-commands:
-  mineresetlite:
-    aliases: [mrl, mr, mine, minereset]
-    description: Manage mines.
